% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loadfrom10X.R
\name{Loadfrom10X}
\alias{Loadfrom10X}
\title{Load Data from 10X Cellranger Output, Optionally Run SoupX, and Create Modified Seurat Object}
\usage{
Loadfrom10X(
  dataDir,
  project = "defaultProject",
  minCells = 3,
  minFeatures = 200,
  useSoupX = FALSE
)
}
\arguments{
\item{dataDir}{Top level CellRanger count output directory.}

\item{project}{Name of the project, used to tag the Seurat object.}

\item{minCells}{Minimum number of cells for a gene to be included in the Seurat object.}

\item{minFeatures}{Minimum number of features for a cell to be included in the Seurat object.}

\item{useSoupX}{Logical, whether to use SoupX to adjust for contamination.}
}
\value{
A Seurat object.
}
\description{
This function reads 10X Cellranger count output from a specified directory
and initializes a Seurat object with the provided parameters. Optionally,
it can use SoupX to estimate and remove cell-free mRNA contamination. After
creating the Seurat object with minimum cells and genes thresholds,
it calculates the percentages of mitochondrial genes (percent.mt),
ribosomal genes (percent.ribo), and hemoglobin genes
(percent.hb) and adds them to the metadata of the Seurat object.
You might also initialize the filtering
}
\examples{
pbmc <- Loadfrom10X(dataDir = "path/to/your/cellranger/outs/folder",
                    project = "ExampleProject",
                    min.cells = 3,
                    min.features = 200,
                    useSoupX = TRUE)
}
