% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneDensity.R
\name{PlotGeneDensity}
\alias{PlotGeneDensity}
\title{Plot Gene Count Density from a Seurat Object}
\usage{
PlotGeneDensity(seuratObject, group = NULL, vlineIntercept = 300)
}
\arguments{
\item{seuratObject}{A Seurat object containing single-cell RNA-seq data.}

\item{vlineIntercept}{The x-intercept for the vertical line in the plot. Defaults to 300.}

\item{sample}{Optional; the name of the metadata column to use for coloring and filling in the plot.
Defaults to orig.ident if not specified.}
}
\value{
A ggplot object showing the density of gene counts per cell.
}
\description{
This function generates a density plot of gene counts per cell from a Seurat object,
allowing for visualization of cell distribution based on the number of detected genes.
The plot uses a logarithmic scale for the x-axis and includes a vertical line as a reference.
}
\examples{
plotGeneDensity(seuratObject)
plotGeneDensity(seuratObject, group = "groups" , vlineIntercept = 300)
}
