% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHemoglobin.R
\name{PlotHemoglobin}
\alias{PlotHemoglobin}
\title{Plot Hemoglobin Gene Percentage}
\usage{
PlotHemoglobin(seurat_obj, group = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA sequencing data.}

\item{group}{The name of the metadata column in the Seurat object that identifies the sample or condition.
This will be used for coloring and filling the density plot. If `NULL`, no grouping is applied.}
}
\value{
A ggplot object representing the density plot of hemoglobin gene percentages.
}
\description{
This function takes a Seurat object and plots the density of hemoglobin gene percentages stored in the object's metadata.
It requires the `Seurat` and `ggplot2` packages. If the hemoglobin gene percentages have not been pre-calculated and stored
in the object using the `PercentageFeatureSet` function with a pattern matching hemoglobin genes (e.g., starting with "Hb"),
this function will first calculate these percentages.
}
\examples{
# Assuming `seurat_obj` is your Seurat object and it has a column 'sample' in its metadata:
plotHemoglobin(seurat_obj, group = "sample")
}
