% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMitochondrial.R
\name{PlotMitochondrial}
\alias{PlotMitochondrial}
\title{Plot Mitochondrial Gene Percentage}
\usage{
PlotMitochondrial(seurat_obj, group = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA sequencing data.}

\item{group}{The name of the metadata column in the Seurat object that identifies the sample or condition.
This will be used for coloring and filling the density plot.}
}
\value{
A ggplot object representing the density plot of mitochondrial gene percentages.
}
\description{
This function takes a Seurat object and plots the density of mitochondrial gene percentages stored in the object's metadata.
It requires the `Seurat` and `ggplot2` packages. If the mitochondrial gene percentages have not been pre-calculated and stored
in the object using the `PercentageFeatureSet` function with a pattern matching mitochondrial genes (e.g., "^MT-"),
this function will first calculate these percentages.
}
\examples{
# Assuming `seurat_obj` is your Seurat object and it has a column 'sample' in its metadata:
plotMitochondrial(seurat_obj, pattern = "(?i)^mt-", group = "groups")
}
