% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNCells.R
\name{PlotNCells}
\alias{PlotNCells}
\title{Plot Number of Cells per Sample from Seurat Object Metadata}
\usage{
PlotNCells(seuratObject, sample = "orig.ident")
}
\arguments{
\item{seuratObject}{A Seurat object with a 'sample' column in the metadata.}

\item{sample}{Optional; the name of the metadata column to use for grouping.}
}
\value{
A ggplot object showing the number of cells per sample.
}
\description{
Plot Number of Cells per Sample from Seurat Object Metadata
}
\examples{
PlotNCellsPerSample(seuratObject)
}
