% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNovelty.R
\name{PlotNoveltyDensity}
\alias{PlotNoveltyDensity}
\title{Plot Novelty Score Density from a Seurat Object}
\usage{
PlotNoveltyDensity(seuratObject, group = NULL, vlineIntercept = 0.8)
}
\arguments{
\item{seuratObject}{A Seurat object containing single-cell RNA-seq data.}

\item{group}{Optional; the name of the metadata column to use for coloring and filling in the plot.
Defaults to orig.ident if not specified.}

\item{vlineIntercept}{The x-intercept for the vertical line in the plot. Defaults to 0.8.}
}
\value{
A ggplot object showing the density of the novelty score.
}
\description{
This function generates a density plot of the novelty score (log10(genes) detected per log10(UMI))
from a Seurat object, allowing for visualization of cell distribution based on this metric.
The plot includes a vertical line as a reference.
}
\examples{
plotNoveltyScoreDensity(seuratObject)
plotNoveltyScoreDensity(seuratObject, group = "sampleGroup", vlineIntercept = 0.8)
}
