% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRibosomal.R
\name{PlotRibosomal}
\alias{PlotRibosomal}
\title{Plot Ribosomal Gene Percentage}
\usage{
PlotRibosomal(seurat_obj, group = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA sequencing data.}

\item{group}{The name of the metadata column in the Seurat object that identifies the sample or condition.
This will be used for coloring and filling the density plot. If `NULL`, no grouping is applied.}
}
\value{
A ggplot object representing the density plot of ribosomal gene percentages.
}
\description{
This function takes a Seurat object and plots the density of ribosomal gene percentages stored in the object's metadata.
It requires the `Seurat` and `ggplot2` packages. If the ribosomal gene percentages have not been pre-calculated and stored
in the object using the `PercentageFeatureSet` function with a pattern matching ribosomal genes (e.g., "^Rp[sl]"),
this function will first calculate these percentages.
}
\examples{
# Assuming `seurat_obj` is your Seurat object and it has a column 'sample' in its metadata:
plotRibosomal(seurat_obj, group = "sample")
}
