% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUMIDensity.R
\name{PlotUMIDensity}
\alias{PlotUMIDensity}
\title{Plot UMI Counts Density from Seurat Object}
\usage{
PlotUMIDensity(seuratObject, sample = NULL)
}
\arguments{
\item{seuratObject}{A Seurat object.}

\item{sample}{Optional; the name of the metadata column to use for coloring.}
}
\value{
A ggplot object showing the density of UMI counts per cell.
}
\description{
Plot UMI Counts Density from Seurat Object
}
\examples{
PlotUMIDensity(seuratObject)
PlotUMIDensity(seuratObject, sample = "sampleColumn")
}
