% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUMIvsGene.R
\name{PlotUMIvsGenes}
\alias{PlotUMIvsGenes}
\title{Plot UMI Counts vs. Number of Genes}
\usage{
PlotUMIvsGenes(seurat_obj, group = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA sequencing data with pre-calculated
`nCount_RNA` and `nFeature_RNA` in its metadata.}

\item{group}{The name of the metadata column in the Seurat object that identifies the grouping
variable for faceting the plot. This allows visualization of data distribution across different groups.}
}
\value{
A ggplot object representing the scatter plot of UMI counts vs. number of genes, faceted by the specified group.
}
\description{
This function takes a Seurat object and generates a scatter plot comparing UMI counts (`nCount_RNA`)
to the number of detected genes (`nFeature_RNA`), with points faceted by a specified grouping variable.
It leverages the `ggplot2` package for plotting. This visualization is useful for assessing the quality
and distribution of single-cell RNA-seq data across different groups.
}
\examples{
# Assuming `seurat_obj` is your Seurat object and it has a column 'group' in its metadata for faceting:
plotUMIvsGenes(seurat_obj, group_col = "group")
}
