% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPCAandElblowPlot.R
\name{RunPCAandElbowPlot}
\alias{RunPCAandElbowPlot}
\title{Run PCA Workflow and Generate Elbow Plot for Seurat Object}
\usage{
RunPCAandElbowPlot(seuratObject)
}
\arguments{
\item{seuratObject}{A Seurat object.}
}
\value{
Generates an elbow plot showing the variance explained by each principal component.
}
\description{
This function takes a Seurat object, normalizes the data, identifies variable features,
scales the data, runs PCA, and generates an elbow plot.
}
